+++ packages/connect/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Breaking
 
-- Removed `createPolkadotJsScClient` from the API, as the `ScProvider` has been upstreamed to `@polkadot/rpc-provider`
+- Removed `createPolkadotJsScClient` from the API, as the `ScProvider` has been upstreamed to `@polkadot/rpc-provider`. ([#909](https://github.com/paritytech/substrate-connect/pull/909))
 
 ### Fixed